--- src/redux/store.js
@@ -1,1 +1,0 @@
-//Redux: State Management Tools
