--- src/utils/constants.js
@@ -1,2 +1,0 @@
-//Constants
-export const API_ROOT = 'http://localhost:8017'
